uptime fix

+++ src/fpsgame/extinfo.h
@@ -130,7 +130,7 @@
         {
             case EXT_UPTIME:
             {
-                putint(p, totalmillis/1000); //in seconds
+                putint(p, uint(totalmillis)/1000); //in seconds
                 break;
             }
